Fixed seg-faults

+++ src/forecast.h
@@ -7,7 +7,7 @@
 
 class Forecast {
     std::vector<std::vector<std::pair<sf::RectangleShape, bool>>> forecast;
-    std::vector<std::pair<std::pair<uint16_t, uint16_t>, std::pair<uint16_t, uint16_t>>> paths;
+    std::vector<std::pair<std::pair<uint64_t, uint64_t>, std::pair<uint64_t, uint64_t>>> paths;
 public:
     Forecast();
     Forecast(const Forecast&);
